pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Template for future projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]